AbstractDatasetInfoBuilder implemented
